Add image_to_image

+++ keras_cv/models/stable_diffusion/stable_diffusion.py
@@ -389,7 +389,7 @@ class StableDiffusionBase:
         input_image = input_image.resize((self.img_width, self.img_height))
         input_image_array = np.array(input_image, dtype=np.float32)[None,...,:3]
 
-        input_image_tensor = ops.cast((input_image_array / 255.0) * 2 - 1, dtype = self.dtype)
+        input_image_tensor = ops.cast((input_image_array / 255.0) * 2 - 1, dtype = np.float32)
         return input_image_tensor
 
     @staticmethod
